Design(DetailVC): LiveActivity 디자인 변경
