+++ SubwayWhenDetailWidget/SubwayWhenDetailWidgetBundle.swift
@@ -1,0 +1,16 @@
+//
+//  SubwayWhenDetailWidgetBundle.swift
+//  SubwayWhenDetailWidget
+//
+//  Created by 이윤수 on 2023/05/09.
+//
+
+import WidgetKit
+import SwiftUI
+
+@main
+struct SubwayWhenDetailWidgetBundle: WidgetBundle {
+    var body: some Widget {
+        SubwayWhenDetailWidgetLiveActivity()
+    }
+}
